requirements.txt
Django==1.5.1
django-grappelli==2.4.4
sorl-thumbnail==11.05.2
psycopg2
Pillow==2.0.0
